Add process name support to `taskkill` command

+++ src/commands/taskkill.js
@@ -1,4 +1,5 @@
 const chalk = require("chalk");
+const psList = require("ps-list");
 
 const { GLOBAL_NAME } = require("../variables/constants");
 
@@ -11,9 +12,35 @@ const InfoMessages = require("../classes/InfoMessages");
 const Verbose = require("../classes/Verbose");
 
 /**
+ * Checks if the user wants to kill the BubbleOS process,
+ * and gives a message if so.
+ *
+ * @param {string | number} pid The PID that the user entered.
+ * @param {boolean} killSelf Whether or not the user allowed BubbleOS to kill itself.
+ * @returns `true` if the user attempted to kill BubbleOS, `false` otherwise.
+ */
+const _killSelfMsg = (pid, killSelf) => {
+  // If the PID is equal to BubbleOS' PID and the user did not give permission to kill itself
+  if (Number(pid) === process.pid && !killSelf) {
+    Verbose.custom("Attempted to kill BubbleOS process...");
+    console.log(
+      chalk.yellow(
+        `You cannot kill the ${GLOBAL_NAME} process. To exit ${GLOBAL_NAME}, run the '${chalk.italic(
+          "exit"
+        )}' command.\nProcess aborted.\n`
+      )
+    );
+
+    return true;
+  }
+
+  return false;
+};
+
+/**
  * Kill a process on the device from the BubbleOS
  * CLI shell, using the provided PID (process
- * identification number).
+ * identification number) or process name.
  *
  * Available arguments:
  * - `-y`: Automatically accept the confirmation
@@ -24,64 +51,64 @@ const Verbose = require("../classes/Verbose");
  * Use this flag at your own risk, as it can remove any
  * unsaved data in BubbleOS!
  *
- * @param {string | number} pid The PID to kill.
+ * @param {string | number} processName The PID or process name to kill.
  * @param {...string} args Arguments to change the behavior of `taskkill` (listed above).
  */
-const taskkill = (pid, ...args) => {
+const taskkill = async (processName, ...args) => {
   try {
     Verbose.initChecker();
-    const pidChk = new Checks(pid);
+    const processChk = new Checks(processName);
 
     Verbose.initArgs();
     const confirm = !args?.includes("-y");
     const silent = args?.includes("-s");
     const killSelf = args?.includes("--kill-self");
 
-    if (pidChk.paramUndefined()) {
+    if (processChk.paramUndefined()) {
       Verbose.chkEmpty();
-      Errors.enterParameter("a PID", "taskkill 1234");
-      return;
-    } else if (isNaN(Number(pid))) {
-      Verbose.custom("PID was detected to not be a number.");
-      Errors.invalidCharacters("PID", "numbers", "numbers/special characters", pid);
-      return;
-    }
-
-    // If the PID is equal to BubbleOS' PID and the user did not give permission to kill itself
-    if (Number(pid) === process.pid && !killSelf) {
-      Verbose.custom("Attempted to kill BubbleOS process...");
-      console.log(
-        chalk.yellow(
-          `You cannot kill the ${GLOBAL_NAME} process. To exit ${GLOBAL_NAME}, run the ${chalk.italic(
-            "'exit'"
-          )} command.\nProcess aborted.\n`
-        )
-      );
+      Errors.enterParameter("a process name or PID", "taskkill test.exe");
       return;
     }
 
     // If the user did not use the '-y' flag
     if (confirm) {
       Verbose.promptUser();
-      if (!_promptForYN(`Are you sure you want to kill the process with PID ${chalk.bold(pid)}?`)) {
+      if (!_promptForYN(`Are you sure you want to kill the process ${chalk.bold(processName)}?`)) {
         console.log(chalk.yellow("Process aborted.\n"));
         return;
       }
     }
 
-    Verbose.custom("Attempting to kill process...");
-    process.kill(Number(pid));
+    if (!isNaN(Number(processName))) {
+      if (_killSelfMsg(processName, killSelf)) return;
+
+      Verbose.custom("Attempting to kill process...");
+      process.kill(Number(processName));
+      return;
+    } else {
+      const processes = await psList();
+      const result = processes.find((obj) => obj.name === processName);
+
+      if (result) {
+        if (_killSelfMsg(result.pid, killSelf)) return;
+
+        Verbose.custom("Attempting to kill process...");
+        process.kill(Number(result.pid));
+        return;
+      }
+    }
 
     // If the user did not request output, show a newline, else, show the success message
-    if (!silent) InfoMessages.success(`Successfully killed the process ${chalk.bold(pid)}.`);
+    if (!silent)
+      InfoMessages.success(`Successfully killed the process ${chalk.bold(processName)}.`);
     else console.log();
   } catch (err) {
     if (err.code === "EPERM") {
       Verbose.permError();
-      Errors.noPermissions("kill the process with PID", pid);
+      Errors.noPermissions("kill the process", processName);
     } else if (err.code === "ESRCH") {
-      Verbose.custom(`The process with PID '${pid}' was detected to not exist.`);
-      Errors.doesNotExist("PID", pid);
+      Verbose.custom(`The process '${processName}' was detected to not exist.`);
+      Errors.doesNotExist("process", processName);
     } else {
       Verbose.fatalError();
       _fatalError(err, false);
